==> R/step6_probe_per_gene_tabulation.R <==
#' step6_probe_per_gene_tabulation
#'
#' This is the step6 function of the TENETR package.
#' This function takes the final optimized enhancer DNA methylation probe to gene
#' links found from step5_optimize_links and tabulates how many of these links are
#' found, both for each gene in the analysis as well as only accepted transcription factor
#' genes from "The Human Transcription Factors" by Lambert et al. 2018.
#' This tabulation is done separately for each of the four hypo or hypermethylated
#' Gplus or Gminus analysis quadrants, as selected by the user.
#'
#' @param TENET_directory Set a path to the TENET directory containing the 'step5' subdirectory and results created by the step5_optimize_links function. This function will also create a new 'step6' subdirectory there containing the results of this function.
#' @param hypermeth_Gplus_analysis Set to TRUE/FALSE depending on if you want to calculate total links by gene/TF for hypermeth probes with G+ links.
#' @param hypermeth_Gminus_analysis Set to TRUE/FALSE depending on if you want to calculate total links by gene/TF for hypermeth probes with G- links.
#' @param hypometh_Gplus_analysis Set to TRUE/FALSE depending on if you want to calculate total links by gene/TF for hypometh probes with G+ links.
#' @param hypometh_Gminus_analysis Set to TRUE/FALSE depending on if you want to calculate total links by gene/TF for hypometh probes with G- links.
#' @return Currently returns tab-delimited "_freq.txt" files that include all genes, or annotated TFs only, organized by decreasing number of linked probes of the given analysis type.
#' @export

step6_probe_per_gene_tabulation <- function(
  TENET_directory,
  hypermeth_Gplus_analysis,
  hypermeth_Gminus_analysis,
  hypometh_Gplus_analysis,
  hypometh_Gminus_analysis
){

  ## Check to make sure at least one analysis type has been selected and return
  ## an error message if at least one hasn't been
  if(
    hypermeth_Gplus_analysis==FALSE &
    hypermeth_Gminus_analysis==FALSE &
    hypometh_Gplus_analysis==FALSE &
    hypometh_Gminus_analysis==FALSE
  ){

    stop(
      "All analysis types have been set to false. Set at least one analysis type to TRUE"
    )
  }

  ## If user has not supplied the final '/' in the TENET directory
  ## add it:
  TENET_directory <- ifelse(
    substring(
      TENET_directory,
      nchar(TENET_directory),
      nchar(TENET_directory)
    ) == '/',
    TENET_directory,
    paste(
      TENET_directory,
      '/',
      sep=''
    )
  )

  ## Create a step6 directory to deposit the output the tabulation files:
  if(
    !dir.exists(
      paste(
        TENET_directory,
        'step6/',
        sep=''
      )
    )
  ){

    dir.create(
      paste(
        TENET_directory,
        'step6/',
        sep=''
      )
    )
  }

  ## Get the dataset of gencode v22 genes:
  gencode_v22_gtf <- TENETR.data::gencode_v22_annotations

  ## Get the gene annotations only:
  gencode_v22_genes <- gencode_v22_gtf[
    gencode_v22_gtf$type=='gene',
  ]

  rownames(gencode_v22_genes) <- sub(
    '\\..*',
    '',
    gencode_v22_genes$gene_id
  )

  ## Get the TF only info:
  TF_dataset <- TENETR.data::human_transcription_factors_dataset

  ## Get the accepted TFs out of the set:
  TF_gene_ID <- TF_dataset[
    TF_dataset$Is.TF.=='Yes',
    'Ensembl.ID'
  ]

  ## Generate results for hypermeth Gplus probes:
  if(hypermeth_Gplus_analysis==TRUE){

    ## Check that the hyper_Gplus_sig_link_zscores_perm_optimized.txt file exists;
    if(
      file.exists(
        paste(
          TENET_directory,
          'step5/',
          'hyper_Gplus_sig_link_zscores_perm_optimized.txt',
          sep=''
        )
      )
    ){

      ## Load the sig_link_zscores_perm_optimized.txt file for hypermeth Gplus
      ## generated by step5 optimize_links function:
      hypermeth_Gplus_sig_links <- read.delim(
        paste(
          TENET_directory,
          'step5/',
          'hyper_Gplus_sig_link_zscores_perm_optimized.txt',
          sep=''
        ),
        sep='\t',
        stringsAsFactors = FALSE
      )

    } else{

      ## Return an error message that the file wasn't found:
      stop('hyper_Gplus_sig_link_zscores_perm_optimized.txt in step5 of TENET directory was not found. Please check that the file exists and consider rerunning the step5 optimize_links function.')
    }

    ## Create a table dataset from this file:
    hypermeth_Gplus_all_gene_dataset <- as.data.frame(
      table(
        hypermeth_Gplus_sig_links$geneID
      ),
      stringsAsFactors=FALSE
    )

    ## Rename the columns:
    colnames(hypermeth_Gplus_all_gene_dataset) <- c(
      'geneID',
      'Freq'
    )

    ## Add the gene names to the dataset:
    hypermeth_Gplus_all_gene_dataset$geneSymbol <- gencode_v22_genes[
      hypermeth_Gplus_all_gene_dataset$geneID,
      'gene_name'
    ]

    ## Sort the data frame by decreasing number of links:
    hypermeth_Gplus_all_gene_dataset <- hypermeth_Gplus_all_gene_dataset[
      order(
        hypermeth_Gplus_all_gene_dataset$Freq,
        decreasing = TRUE
      ),
    ]

    ## Save that dataset as the all genes dataset:
    write.table(
      hypermeth_Gplus_all_gene_dataset,
      file= paste(
        TENET_directory,
        'step6/',
        'hyper_Gplus_links_all_gene_freq.txt',
        sep=''
      ),
      quote= FALSE,
      sep='\t',
      row.names = FALSE
    )

    ## Now get only the TFs from this table:
    hypermeth_Gplus_TF_gene_dataset <- hypermeth_Gplus_all_gene_dataset[
      hypermeth_Gplus_all_gene_dataset$geneID %in% TF_gene_ID,
    ]

    ## Save that dataset as TF only dataset:
    write.table(
      hypermeth_Gplus_TF_gene_dataset,
      file= paste(
        TENET_directory,
        'step6/',
        'hyper_Gplus_links_all_TF_freq.txt',
        sep=''
      ),
      quote= FALSE,
      sep='\t',
      row.names = FALSE
    )
  }

  ## Generate results for hypermeth Gplus probes:
  if(hypermeth_Gminus_analysis==TRUE){

    ## Check that the hyper_Gminus_sig_link_zscores_perm_optimized.txt file exists;
    if(
      file.exists(
        paste(
          TENET_directory,
          'step5/',
          'hyper_Gminus_sig_link_zscores_perm_optimized.txt',
          sep=''
        )
      )
    ){

      ## Load the sig_link_zscores_perm_optimized.txt file for hypermeth Gminus
      ## generated by step5 optimize_links function:
      hypermeth_Gminus_sig_links <- read.delim(
        paste(
          TENET_directory,
          'step5/',
          'hyper_Gminus_sig_link_zscores_perm_optimized.txt',
          sep=''
        ),
        sep='\t',
        stringsAsFactors = FALSE
      )

    } else{

      ## Return an error message that the file wasn't found:
      stop('hyper_Gminus_sig_link_zscores_perm_optimized.txt in step5 of TENET directory was not found. Please check that the file exists and consider rerunning the step5 optimize_links function.')
    }

    ## Create a table dataset from this file:
    hypermeth_Gminus_all_gene_dataset <- as.data.frame(
      table(
        hypermeth_Gminus_sig_links$geneID
      ),
      stringsAsFactors=FALSE
    )

    ## Rename the columns:
    colnames(hypermeth_Gminus_all_gene_dataset) <- c(
      'geneID',
      'Freq'
    )

    ## Add the gene names to the dataset:
    hypermeth_Gminus_all_gene_dataset$geneSymbol <- gencode_v22_genes[
      hypermeth_Gminus_all_gene_dataset$geneID,
      'gene_name'
    ]

    ## Sort the data frame by decreasing number of links:
    hypermeth_Gminus_all_gene_dataset <- hypermeth_Gminus_all_gene_dataset[
      order(
        hypermeth_Gminus_all_gene_dataset$Freq,
        decreasing = TRUE
      ),
    ]

    ## Save that dataset as the all genes dataset:
    write.table(
      hypermeth_Gminus_all_gene_dataset,
      file= paste(
        TENET_directory,
        'step6/',
        'hyper_Gminus_links_all_gene_freq.txt',
        sep=''
      ),
      quote= FALSE,
      sep='\t',
      row.names = FALSE
    )

    ## Now get only the TFs from this table:
    hypermeth_Gminus_TF_gene_dataset <- hypermeth_Gminus_all_gene_dataset[
      hypermeth_Gminus_all_gene_dataset$geneID %in% TF_gene_ID,
    ]

    ## Save that dataset as TF only dataset:
    write.table(
      hypermeth_Gminus_TF_gene_dataset,
      file= paste(
        TENET_directory,
        'step6/',
        'hyper_Gminus_links_all_TF_freq.txt',
        sep=''
      ),
      quote= FALSE,
      sep='\t',
      row.names = FALSE
    )
  }

  ## Generate results for hypometh Gplus probes:
  if(hypometh_Gplus_analysis==TRUE){

    ## Check that the hypo_Gplus_sig_link_zscores_perm_optimized.txt file exists;
    if(
      file.exists(
        paste(
          TENET_directory,
          'step5/',
          'hypo_Gplus_sig_link_zscores_perm_optimized.txt',
          sep=''
        )
      )
    ){

      ## Load the sig_link_zscores_perm_optimized.txt file for hypometh Gplus
      ## generated by step5 optimize_links function:
      hypometh_Gplus_sig_links <- read.delim(
        paste(
          TENET_directory,
          'step5/',
          'hypo_Gplus_sig_link_zscores_perm_optimized.txt',
          sep=''
        ),
        sep='\t',
        stringsAsFactors = FALSE
      )

    } else{

      ## Return an error message that the file wasn't found:
      stop('hypo_Gplus_sig_link_zscores_perm_optimized.txt in step5 of TENET directory was not found. Please check that the file exists and consider rerunning the step5 optimize_links function.')
    }

    ## Create a table dataset from this file:
    hypometh_Gplus_all_gene_dataset <- as.data.frame(
      table(
        hypometh_Gplus_sig_links$geneID
      ),
      stringsAsFactors=FALSE
    )

    ## Rename the columns:
    colnames(hypometh_Gplus_all_gene_dataset) <- c(
      'geneID',
      'Freq'
    )

    ## Add the gene names to the dataset:
    hypometh_Gplus_all_gene_dataset$geneSymbol <- gencode_v22_genes[
      hypometh_Gplus_all_gene_dataset$geneID,
      'gene_name'
    ]

    ## Sort the data frame by decreasing number of links:
    hypometh_Gplus_all_gene_dataset <- hypometh_Gplus_all_gene_dataset[
      order(
        hypometh_Gplus_all_gene_dataset$Freq,
        decreasing = TRUE
      ),
    ]

    ## Save that dataset as the all genes dataset:
    write.table(
      hypometh_Gplus_all_gene_dataset,
      file= paste(
        TENET_directory,
        'step6/',
        'hypo_Gplus_links_all_gene_freq.txt',
        sep=''
      ),
      quote= FALSE,
      sep='\t',
      row.names = FALSE
    )

    ## Now get only the TFs from this table:
    hypometh_Gplus_TF_gene_dataset <- hypometh_Gplus_all_gene_dataset[
      hypometh_Gplus_all_gene_dataset$geneID %in% TF_gene_ID,
    ]

    ## Save that dataset as TF only dataset:
    write.table(
      hypometh_Gplus_TF_gene_dataset,
      file= paste(
        TENET_directory,
        'step6/',
        'hypo_Gplus_links_all_TF_freq.txt',
        sep=''
      ),
      quote= FALSE,
      sep='\t',
      row.names = FALSE
    )
  }

  ## Generate results for hypometh Gplus probes:
  if(hypometh_Gminus_analysis==TRUE){

    ## Check that the hypo_Gminus_sig_link_zscores_perm_optimized.txt file exists;
    if(
      file.exists(
        paste(
          TENET_directory,
          'step5/',
          'hypo_Gminus_sig_link_zscores_perm_optimized.txt',
          sep=''
        )
      )
    ){

      ## Load the sig_link_zscores_perm_optimized.txt file for hypometh Gminus
      ## generated by step5 optimize_links function:
      hypometh_Gminus_sig_links <- read.delim(
        paste(
          TENET_directory,
          'step5/',
          'hypo_Gminus_sig_link_zscores_perm_optimized.txt',
          sep=''
        ),
        sep='\t',
        stringsAsFactors = FALSE
      )

    } else{

      ## Return an error message that the file wasn't found:
      stop('hypo_Gminus_sig_link_zscores_perm_optimized.txt in step5 of TENET directory was not found. Please check that the file exists and consider rerunning the step5 optimize_links function.')
    }

    ## Create a table dataset from this file:
    hypometh_Gminus_all_gene_dataset <- as.data.frame(
      table(
        hypometh_Gminus_sig_links$geneID
      ),
      stringsAsFactors=FALSE
    )

    ## Rename the columns:
    colnames(hypometh_Gminus_all_gene_dataset) <- c(
      'geneID',
      'Freq'
    )

    ## Add the gene names to the dataset:
    hypometh_Gminus_all_gene_dataset$geneSymbol <- gencode_v22_genes[
      hypometh_Gminus_all_gene_dataset$geneID,
      'gene_name'
    ]

    ## Sort the data frame by decreasing number of links:
    hypometh_Gminus_all_gene_dataset <- hypometh_Gminus_all_gene_dataset[
      order(
        hypometh_Gminus_all_gene_dataset$Freq,
        decreasing = TRUE
      ),
    ]

    ## Save that dataset as the all genes dataset:
    write.table(
      hypometh_Gminus_all_gene_dataset,
      file= paste(
        TENET_directory,
        'step6/',
        'hypo_Gminus_links_all_gene_freq.txt',
        sep=''
      ),
      quote= FALSE,
      sep='\t',
      row.names = FALSE
    )

    ## Now get only the TFs from this table:
    hypometh_Gminus_TF_gene_dataset <- hypometh_Gminus_all_gene_dataset[
      hypometh_Gminus_all_gene_dataset$geneID %in% TF_gene_ID,
    ]

    ## Save that dataset as TF only dataset:
    write.table(
      hypometh_Gminus_TF_gene_dataset,
      file= paste(
        TENET_directory,
        'step6/',
        'hypo_Gminus_links_all_TF_freq.txt',
        sep=''
      ),
      quote= FALSE,
      sep='\t',
      row.names = FALSE
    )
  }
}
